qrexec-lib: always reject non-UTF8 paths
Various GLib functions assume that inputs are valid UTF-8 but don't
check that.  Unlike Rust, GLib doesn't have any static checks for this.
Since almost all files have names that are valid UTF-8, bugs in handling
of filenames that are not valid UTF-8 are unlikely to be found by
testing.  Therefore, accepting filenames that are not valid UTF-8 is
quite risky.  Reject them unconditionally.